Btc.Address behaviour (btc v0.1.0)

Bitcoin P2PK, P2PKH, P2SH, P2WPKH, P2WSH script/address generate and validate demo.

Supported

P2PKH

P2SH

P2WPKH

P2WSH

Usage

entropy = Btc.Util.strong_rand()
{:ok, {priv_key, address}} = Btc.Address.generate(:p2pkh, :mainnet, entropy)
true = Btc.Address.verify?(:p2pkh, :mainnet, address)

{:ok, {priv_key, address}} = Btc.Address.generate(:p2sh, :mainnet)
true = Btc.Address.verify?(:p2sh, :mainnet, address)

{:ok, {priv_key, address}} = Btc.Address.generate(:p2wpkh, :testnet)
true = Btc.Address.verify?(:p2wpkh, :testnet, address)

Types

address()

Specs

address() :: String.t()

compression()

Specs

compression() :: :compressed | :uncompressed

entropy()

Specs

entropy() :: String.t()

generate_result()

Specs

generate_result() ::
  {:ok, {private_key :: private_key(), address :: address()}} | {:error, atom()}

network()

Specs

network() :: :mainnet | :testnet

https://en.bitcoin.it/wiki/Testnet

private_key()

Specs

private_key() :: <<_::256>>

public_key()

Specs

public_key() :: String.t()

type()

Specs

type() :: :p2wsh | :p2wpkh | :p2sh | :p2pkh

Functions

create_public_key(private_key, compression \\ :compressed)

Specs

create_public_key(private_key :: private_key(), compression :: compression()) ::
  {:ok, public_key()} | {:error, term()}

https://github.com/mbrix/libsecp256k1/blob/master/test/libsecp256k1_tests.erl

generate(type, network, entropy \\ Util.strong_rand())

Specs

generate(type :: type(), network :: network(), entropy :: entropy()) ::
  generate_result()

generate_private_key(entropy)

Specs

generate_private_key(entropy :: String.t()) ::
  {:ok, private_key :: private_key()}

valid?(type, network, address)

Specs

valid?(type :: type(), network :: network(), address :: String.t()) :: boolean()

Callbacks

address_valid?(address, network)

Specs

address_valid?(address :: address(), network :: network()) :: boolean()

Given a address and a network type, check if the address is valid

generate_address(entropy, network)

Specs

generate_address(entropy :: entropy(), network :: network()) ::
  generate_result()

Given a 32 bytes random binary and a network type, return a valid bitcoin address.

type()

Specs

type() :: type()

Return type of current module