Btc.Addresses.P2sh (btc v0.1.0)

Pay-to-Script-Hash

Links

https://en.bitcoin.it/wiki/Pay_to_script_hash

https://en.bitcoin.it/wiki/BIP_0016

https://en.bitcoin.it/wiki/List_of_address_prefixes

https://en.bitcoin.it/wiki/Script#Opcodes

ScriptPubKey and ScriptSig Display

scriptSig: [signature] {[pubkey] OP_CHECKSIG}
scriptPubKey: OP_HASH160 [20-byte-hash of {[pubkey] OP_CHECKSIG} ] OP_EQUAL

Struct

A 1 of 1 multisig transaction is created by concatenating the following byte values:

OP_0 0x14 HASH160( K )

A = BASE58CHECK( 0x05 HASH160( OP_0 0x14 HASH160( K ) ) )
Where 0x05 is the version byte for script hash addresses and 0x14 is
the amount of bytes to expect in the HASH160( K ).

Summary

mainnet

prefix: 5

leading symbol: 3

example: 3EktnHQD7RiAE6uzMj2ZifT9YgRrkSgzQX

testnet

prefix: 196

leading symbol: 2

example: 2MzQwSSnBHWHqSAqtTVQ6v47XtaisrJa1Vc

Doctest mainnet

iex> network = :mainnet
...> entropy = Btc.Util.strong_rand()
...> {:ok, {_priv_key, address}} = Elixir.Btc.Addresses.P2sh.generate_address(entropy, network)
...> Elixir.Btc.Addresses.P2sh.address_valid?(address, network)
true

Doctest testnet

iex> network = :testnet
...> entropy = Btc.Util.strong_rand()
...> {:ok, {_priv_key, address}} = Elixir.Btc.Addresses.P2sh.generate_address(entropy, network)
...> Elixir.Btc.Addresses.P2sh.address_valid?(address, network)
true

Functions

address_valid?(address, network)

Example

iex> Elixir.Btc.Addresses.P2sh.address_valid?("3EktnHQD7RiAE6uzMj2ZifT9YgRrkSgzQX", :mainnet)
true
iex> Elixir.Btc.Addresses.P2sh.address_valid?("2MzQwSSnBHWHqSAqtTVQ6v47XtaisrJa1Vc", :testnet)
true
iex> Elixir.Btc.Addresses.P2sh.address_valid?("foo", :mainnet)
false