Btc.Bip0173 (btc v0.1.0)

Summary

Bech32 is a segwit address format specified by BIP 0173. This
address format is also known as "bc1 addresses". Bech32 is more
efficient with block space. As of October 2020, the Bech32
address format is supported in many popular wallets and is the
preferred address scheme

A Bech32[2] string is at most 90 characters long and consists of:

The human-readable part, which is intended to convey the type of data, or anything else that is relevant to the reader. This part MUST contain 1 to 83 US-ASCII characters, with each character having a value in the range [33-126]. HRP validity may be further restricted by specific applications.

The separator, which is always "1". In case "1" is allowed inside the human-readable part, the last one in the string is the separator[3].

The data part, which is at least 6 characters long and only consists of alphanumeric characters excluding "1", "b", "i", and "o"[4].

Example

Examples of the address format being used on mainnet are the TXIDs
4ef47f6eb681d5d9fa2f7e16336cd629303c635e8da51e425b76088be9c8744c
514a33f1d46179b89e1fea7bbb07b682ab14083a276979f91038369d1a8d689b
And addresses
bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq
bc1qc7slrfxkknqcq2jevvvkdgvrt8080852dfjewde450xdlk4ugp7szw5tk9

Functions

decode(data)

Specs

decode(b :: String.t()) ::
  {:ok, network :: Btc.Address.network(), witness_version :: integer(),
   data :: String.t()}
  | {:error, reason :: atom()}

Software interpreting a segwit address:

MUST verify that the human-readable part is "bc" for mainnet and "tb" for testnet.

MUST verify that the first decoded data value (the witness version) is between 0 and 16, inclusive.

https://github.com/f2pool/bech32-elixir/blob/master/lib/bech32.ex#L257 https://en.bitcoin.it/wiki/Bech32 https://en.bitcoin.it/wiki/BIP_0173

Example

iex> match?({:ok, :mainnet, 3, _}, Elixir.Btc.Bip0173.decode("bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4"))
true

encode(hrp, data)

See Bech32.encode/2.